is.Learning.DDD-main/e-commerce vs/requirements.txt
Flask>=2.3
mysql-connector-python>=8.0

